Extrect other parts of ManagmentActor
