+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/NussknackerApp.scala
@@ -129,7 +129,7 @@ trait NusskanckerDefaultAppRouter extends NusskanckerAppRouter {
 
     val scenarioResolver = new ScenarioResolver(subprocessResolver)
     val actionRepository = DbProcessActionRepository.create(dbConfig, modelData)
-    deploymentService = new DeploymentService(processRepository, actionRepository, scenarioResolver)
+    deploymentService = new DeploymentService(processRepository, actionRepository, scenarioResolver, processChangeListener)
     reload.init() // we need to init processing type data after deployment service creation to make sure that it will be done using correct classloader and that won't cause further delays during handling requests
     val processActivityRepository = new DbProcessActivityRepository(dbConfig)
 
@@ -143,7 +143,7 @@ trait NusskanckerDefaultAppRouter extends NusskanckerAppRouter {
     val newProcessPreparer = NewProcessPreparer(typeToConfig, additionalProperties)
 
     val systemRequestTimeout = system.settings.config.getDuration("akka.http.server.request-timeout")
-    val managementActor = system.actorOf(ManagementActor.props(managers, processRepository, actionRepository, scenarioResolver, processChangeListener, deploymentService), "management")
+    val managementActor = system.actorOf(ManagementActor.props(managers, processRepository, scenarioResolver, deploymentService), "management")
     val processService = new DBProcessService(managementActor, systemRequestTimeout, newProcessPreparer,
       processCategoryService, processResolving, dbRepositoryManager, processRepository, actionRepository,
       writeProcessRepository
