+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/DeploymentService.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.ui.process.deployment
 
+import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
@@ -8,10 +9,12 @@ import pl.touk.nussknacker.engine.canonize.ProcessCanonizer
 import pl.touk.nussknacker.engine.deployment.{DeploymentData, DeploymentId, User}
 import pl.touk.nussknacker.restmodel.process.{ProcessIdWithName, ProcessingType}
 import pl.touk.nussknacker.restmodel.processdetails.BaseProcessDetails
+import pl.touk.nussknacker.ui.api.ListenerApiUser
 import pl.touk.nussknacker.ui.db.entity.ProcessActionEntityData
-import pl.touk.nussknacker.ui.process.repository.DeploymentComment
+import pl.touk.nussknacker.ui.listener.ProcessChangeEvent.{OnDeployActionFailed, OnDeployActionSuccess, OnFinished}
+import pl.touk.nussknacker.ui.listener.{ProcessChangeListener, User => ListenerUser}
 import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository.ProcessNotFoundError
-import pl.touk.nussknacker.ui.process.repository.{DbProcessActionRepository, FetchingProcessRepository}
+import pl.touk.nussknacker.ui.process.repository.{DbProcessActionRepository, DeploymentComment, FetchingProcessRepository}
 import pl.touk.nussknacker.ui.security.api.{LoggedUser, NussknackerInternalUser}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -23,15 +26,18 @@ import scala.util.{Failure, Success}
   */
 class DeploymentService(processRepository: FetchingProcessRepository[Future],
                         actionRepository: DbProcessActionRepository,
-                        scenarioResolver: ScenarioResolver)(implicit val ec: ExecutionContext) {
+                        scenarioResolver: ScenarioResolver,
+                        processChangeListener: ProcessChangeListener)(implicit val ec: ExecutionContext) extends LazyLogging {
 
   def cancelProcess(processId: ProcessIdWithName, deploymentComment: Option[DeploymentComment], performCancel: ProcessIdWithName => Future[Unit])(implicit user: LoggedUser): Future[ProcessActionEntityData] = {
-    for {
-      _ <- performCancel(processId)
-      maybeVersion <- findDeployedVersion(processId)
-      version <- processDataExistOrFail(maybeVersion, processId.name.value)
-      result <- actionRepository.markProcessAsCancelled(processId.id, version, deploymentComment)
-    } yield result
+    withDeploymentActionNotification(processId, "cancel", deploymentComment) {
+      for {
+        _ <- performCancel(processId)
+        maybeVersion <- findDeployedVersion(processId)
+        version <- processDataExistOrFail(maybeVersion, processId.id)
+        result <- actionRepository.markProcessAsCancelled(processId.id, version, deploymentComment)
+      } yield result
+    }
   }
 
   def getDeployedScenarios(processingType: ProcessingType): Future[List[DeployedScenarioData]] = {
@@ -58,22 +64,42 @@ class DeploymentService(processRepository: FetchingProcessRepository[Future],
   }
 
   //inner Future in result allows to wait for deployment finish, while outer handles validation
-  def deployProcess(processId: ProcessId,
+  def deployProcess(processIdWithName: ProcessIdWithName,
                     savepointPath: Option[String],
                     deploymentComment: Option[DeploymentComment],
                     deploymentManager: ProcessingType => DeploymentManager)
                    (implicit user: LoggedUser): Future[Future[ProcessActionEntityData]] = {
     for {
-      maybeProcess <- processRepository.fetchLatestProcessDetailsForProcessId[CanonicalProcess](processId)
-      process <- processDataExistOrFail(maybeProcess, processId.value.toString)
-      result <- deployAndSaveProcess(process, savepointPath, deploymentComment, deploymentManager(process.processingType))
+        maybeProcess <- processRepository.fetchLatestProcessDetailsForProcessId[CanonicalProcess](processIdWithName.id)
+        process <- processDataExistOrFail(maybeProcess, processIdWithName.id)
+        result <- deployAndSaveProcess(process, savepointPath, deploymentComment, deploymentManager(process.processingType))
     } yield result
   }
 
-  private def processDataExistOrFail[T](maybeProcess: Option[T], processId: String): Future[T] = {
+  //TODO: there is small problem here: if no one invokes process status for long time, Flink can remove process from history
+  //- then it's gone, not finished.
+  def handleFinishedProcess(idWithName: ProcessIdWithName, processState: Option[ProcessState]): Future[Unit] = {
+    implicit val user: NussknackerInternalUser.type = NussknackerInternalUser
+    implicit val listenerUser: ListenerUser = ListenerApiUser(user)
+    processState match {
+      case Some(state) if state.status.isFinished =>
+        findDeployedVersion(idWithName).flatMap {
+          case Some(version) => {
+            val finishedDeploymentComment = DeploymentComment.unsafe("Scenario finished")
+            actionRepository.markProcessAsCancelled(idWithName.id, version, Some(finishedDeploymentComment)).map(_ =>
+              processChangeListener.handle(OnFinished(idWithName.id, version))
+            )
+          }
+          case _ => Future.successful(())
+        }
+      case _ => Future.successful(())
+    }
+  }
+
+  private def processDataExistOrFail[T](maybeProcess: Option[T], processId: ProcessId): Future[T] = {
     maybeProcess match {
       case Some(processData) => Future.successful(processData)
-      case None => Future.failed(ProcessNotFoundError(processId))
+      case None => Future.failed(ProcessNotFoundError(processId.value.toString))
     }
   }
 
@@ -89,11 +115,14 @@ class DeploymentService(processRepository: FetchingProcessRepository[Future],
     } yield (resolvedCanonicalProces, deploymentData)
 
     validatedData.map { case (resolvedCanonicalProces, deploymentData) =>
-      for {
-        _ <- deploymentManager.deploy(processVersion, deploymentData, resolvedCanonicalProces, savepointPath)
-        deployedActionData <- actionRepository.markProcessAsDeployed(
-          process.processId, process.processVersionId, process.processingType, deploymentComment)
-      } yield deployedActionData
+      //we notify of deployment finish/fail only if initial validation succeeded
+      withDeploymentActionNotification(process.idWithName, "deploy", deploymentComment) {
+        for {
+          _ <- deploymentManager.deploy(processVersion, deploymentData, resolvedCanonicalProces, savepointPath)
+          deployedActionData <- actionRepository.markProcessAsDeployed(
+            process.processId, process.processVersionId, process.processingType, deploymentComment)
+        } yield deployedActionData
+      }
     }
   }
 
@@ -107,5 +136,21 @@ class DeploymentService(processRepository: FetchingProcessRepository[Future],
   } yield lastAction.map(la => la.processVersionId)
 
   private def toManagerUser(loggedUser: LoggedUser) = User(loggedUser.id, loggedUser.username)
+
+  private def withDeploymentActionNotification(processIdWithName: ProcessIdWithName,
+                               actionName: String,
+                               deploymentComment: Option[DeploymentComment])(action: => Future[ProcessActionEntityData])(implicit user: LoggedUser): Future[ProcessActionEntityData] = {
+    implicit val listenerUser: ListenerUser = ListenerApiUser(user)
+    val actionToRun = action
+    actionToRun.onComplete {
+      case Failure(failure) =>
+        logger.error(s"Action: $actionName of ${processIdWithName.name} finished with failure", failure)
+        processChangeListener.handle(OnDeployActionFailed(processIdWithName.id, failure))
+      case Success(details) =>
+        logger.info(s"Finishing $actionName of ${processIdWithName.name}")
+        processChangeListener.handle(OnDeployActionSuccess(details.processId, details.processVersionId, deploymentComment, details.performedAtTime, details.action))
+    }
+    actionToRun
+  }
 
 }
